+++ tests/test_utils.py
@@ -6,7 +6,6 @@
 # Copyright (C) 2010, Manfred Moitzi
 # License: MIT License
 
-import sys
 import unittest
 
 from svgwrite.utils import rgb, AutoID, get_unit